# openapi.json
{
  "components": {
    "schemas": {
...
      "AnchorProofEntry": {
        "description": "One audit row of this batch, and the proof for it if one exists yet.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "anchor_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnchorTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "anchored_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchored At"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "integer"
          },
          "merkle_path": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProofStepOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merkle Path"
          },
          "merkle_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merkle Root"
          },
          "ots_proof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ots Proof"
          },
          "row_hash": {
            "title": "Row Hash",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "anchored",
              "confirmed"
            ],
            "title": "Status",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "audit_id",
          "action",
          "row_hash",
          "status"
        ],
        "title": "AnchorProofEntry",
        "type": "object"
      },
      "AnchorProofResponse": {
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AnchorProofEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "partial",
              "anchored",
              "confirmed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "batch_id",
          "public_id",
          "status",
          "entries"
        ],
        "title": "AnchorProofResponse",
        "type": "object"
      },
      "AnchorTarget": {
        "description": "Where a Merkle root is published (04 \u00a75.2).\n\nThe target is settled: OpenTimestamps to Bitcoin (05 \u00a76.3). `polygon` is a\nreserved fallback (06 \u00a79) \u2014 the value exists so switching would not need a\nmigration; no code path in v2 consumes it.",
        "enum": [
          "opentimestamps",
          "polygon"
        ],
        "title": "AnchorTarget",
        "type": "string"
      },
...
      "BatchVerifyResponse": {
        "properties": {
          "anchor_status": {
            "enum": [
              "pending",
              "partial",
              "anchored",
              "confirmed"
            ],
            "title": "Anchor Status",
            "type": "string"
          },
          "batch_id": {
            "title": "Batch Id",
            "type": "integer"
          },
          "conformance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConformanceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "verification": {
            "$ref": "#/components/schemas/StageVerifications"
          }
        },
        "required": [
          "batch_id",
          "public_id",
          "state",
          "metadata",
          "verification",
          "conformance",
          "anchor_status"
        ],
        "title": "BatchVerifyResponse",
        "type": "object"
      },
...
      "ConformanceOut": {
        "description": "The Codex/KEBS verdict re-derived from the lab row now, under the rule\nset recorded with it. Every limit and its basis travel with the result, so\na reader never needs to know which rules were in force.",
        "properties": {
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/ConformanceParameterOut"
            },
            "title": "Parameters",
            "type": "array"
          },
          "rule_set_version": {
            "title": "Rule Set Version",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "pass",
              "fail",
              "incomplete"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "rule_set_version",
          "verdict",
          "parameters"
        ],
        "title": "ConformanceOut",
        "type": "object"
      },
      "ConformanceParameterOut": {
        "properties": {
          "basis": {
            "title": "Basis",
            "type": "string"
          },
          "comparator": {
            "enum": [
              "<=",
              ">="
            ],
            "title": "Comparator",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "string"
          },
          "measured": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "status": {
            "enum": [
              "pass",
              "fail",
              "not_measured"
            ],
            "title": "Status",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "parameter",
          "measured",
          "unit",
          "comparator",
          "limit",
          "basis",
          "status"
        ],
        "title": "ConformanceParameterOut",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "OpenAPI shape of the error envelope, so documented error responses match\nwhat the handler emits (keeps the Schemathesis contract test honest).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetadataPublic": {
        "description": "The farmer's declaration as the anonymous `/verify` view shows it.\n\nNo `notes`: free text can carry personal data whatever it was meant for,\nand this view is broadcast to anyone who scans a jar.",
        "properties": {
          "apiary_management_method": {
            "title": "Apiary Management Method",
            "type": "string"
          },
          "expected_yield_kg": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Expected Yield Kg",
            "type": "string"
          },
          "harvest_window_end": {
            "format": "date",
            "title": "Harvest Window End",
            "type": "string"
          },
          "harvest_window_start": {
            "format": "date",
            "title": "Harvest Window Start",
            "type": "string"
          },
          "honey_type": {
            "title": "Honey Type",
            "type": "string"
          }
        },
        "required": [
          "honey_type",
          "expected_yield_kg",
          "harvest_window_start",
          "harvest_window_end",
          "apiary_management_method"
        ],
        "title": "MetadataPublic",
        "type": "object"
      },
...
      "ProofStepOut": {
        "description": "One sibling on the path to the root; `position` is the sibling's side.",
        "properties": {
          "position": {
            "enum": [
              "left",
              "right"
            ],
            "title": "Position",
            "type": "string"
          },
          "sibling": {
            "title": "Sibling",
            "type": "string"
          }
        },
        "required": [
          "sibling",
          "position"
        ],
        "title": "ProofStepOut",
        "type": "object"
      },
...
      "StageVerificationOut": {
        "description": "One stage of the three-way match.\n\n`payload` is the canonical pre-image as the stage row stands now. When\n`payload_precision` is `exact`, hashing it reproduces `recomputed_hash`.\nWhen it is `reduced`, the fields in `redacted_fields` were coarsened or\nwithheld for privacy and **both hashes are omitted**, because they commit\nto the exact values; `match` is then the server's own comparison.",
        "properties": {
          "anchor_status": {
            "anyOf": [
              {
                "enum": [
                  "pending",
                  "anchored",
                  "confirmed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Status"
          },
          "audit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Id"
          },
          "match": {
            "title": "Match",
            "type": "boolean"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "payload_precision": {
            "enum": [
              "exact",
              "reduced"
            ],
            "title": "Payload Precision",
            "type": "string"
          },
          "recomputed_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recomputed Hash"
          },
          "recorded_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded Hash"
          },
          "redacted_fields": {
            "items": {
              "type": "string"
            },
            "title": "Redacted Fields",
            "type": "array"
          }
        },
        "required": [
          "audit_id",
          "payload",
          "recomputed_hash",
          "recorded_hash",
          "match",
          "anchor_status",
          "payload_precision",
          "redacted_fields"
        ],
        "title": "StageVerificationOut",
        "type": "object"
      },
      "StageVerifications": {
        "description": "`null` means the batch has not reached that stage yet.",
        "properties": {
          "apiary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "distribution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "harvest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "lab": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "packaging": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "process": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageVerificationOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "apiary",
          "metadata",
          "harvest",
          "process",
          "lab",
          "packaging",
          "distribution"
        ],
        "title": "StageVerifications",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Honey provenance and traceability: hash-chained audit log, Merkle anchoring via OpenTimestamps, PostGIS geospatial records.",
    "title": "ApiChain v2 Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/public/batches/{public_id}/anchor-proof": {
      "get": {
        "description": "Everything needed to verify this batch's records offline.\n\nOne entry per audit row, each carrying its own status, so a record that\nexists in our log but has no public anchor yet is reported as `pending`\nrather than as a missing field (03 \u00a76).",
        "operationId": "batch_anchor_proof_v2_public_batches__public_id__anchor_proof_get",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 32,
              "pattern": "^[0-9a-f]{32}$",
              "title": "Public Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnchorProofResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Anchor Proof",
        "tags": [
          "public"
        ]
      }
    },
    "/v2/public/batches/{public_id}/verify": {
      "get": {
        "description": "The three-way match over every recorded stage of this batch.\n\nBeing anonymous, it applies `verification.FIELD_POLICY` (hive coordinates\nat 2 dp, altitude at 100 m, free text and the analyst's name withheld) and\npublishes no hashes for a redacted block. Staff see exact values in an\nauthenticated view, never here. Reports facts only: no score, no band \u2014\nthose mappings are the client's (03 \u00a71, \u00a76).",
        "operationId": "verify_batch_v2_public_batches__public_id__verify_get",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 32,
              "pattern": "^[0-9a-f]{32}$",
              "title": "Public Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Batch",
        "tags": [
          "public"
        ]
      }
    }
  }
}